Update mongodb, routes and upload library
