+++ .meteor/versions
@@ -114,7 +114,7 @@ summernote:standalone@0.8.1
 templating@1.2.14
 templating-tools@1.1.1
 todda00:friendly-slugs@0.4.0
-tomi:upload-jquery@2.2.1
+tomi:upload-jquery@2.4.0
 tomi:upload-server@1.3.3
 tracker@1.2.0
 udondan:slick@1.3.11_1
