+++ imports/api/films/films.js
@@ -1,4 +1,6 @@
-Films = new Mongo.Collection("films");
+import { Mongo } from 'meteor/mongo';
+
+export const Films = new Mongo.Collection("films");
 
 Films.friendlySlugs({
   slugFrom: 'title',
